+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Validate YAML in code editor WebUI.
 
+- Exposing membership options with argparse
+
 ### Changed
 
 - Network error shows with fixed splash panel instead of notification.